+++ models/pack.py
@@ -3,6 +3,6 @@ from pydantic import BaseModel
 
 class PackModel(BaseModel):
     name: str
-    size: tuple[int, int, int]
+    size: tuple[float, float, float]
     material: str
     image_url: str
